requirements.txt
flask>=3.1.2
flask-sqlalchemy>=3.1.1
flask-socketio>=5.5.1
flask-cors>=4.0.0
openpyxl>=3.1.5
pandas>=2.3.3
pillow>=12.0.0
python-socketio>=5.14.2
qrcode>=8.2
python-dotenv>=1.0.0
gunicorn>=21.2.0
reportlab>=4.0.0
apscheduler>=3.10.4
pymysql>=1.1.0
requests>=2.31.0
werkzeug>=3.0.0
python-engineio>=4.8.0
SQLAlchemy>=2.0.0
Jinja2>=3.1.2
MarkupSafe>=2.1.1